renamed settings sub-folders and sub-components
