+++ src/components/pages/Settings/SchoolSettings/index.tsx
@@ -1,20 +1,20 @@
-import { BoldText } from '~reusables/ui/Text';
-import SchoolLogoSetting from './LogoSettings/index';
-import SchoolNameSetting from './NameSettings';
-import AcademicTerm from '~components/pages/Settings/SchoolSettings/TermSettings';
 import { List } from '~components/reusables/List/List';
+import { BoldText } from '~reusables/ui/Text';
+import LogoSetting from './LogoSetting/index';
+import NameSetting from './NameSetting';
+import TermSetting from './TermSetting/index';
 
-const SchoolSettings = () => (
+const SchoolSetting = () => (
   <div>
     <BoldText className="py-6 border-b border-gray-100">
       School Settings
     </BoldText>
     <List>
-      <SchoolNameSetting />
-      <SchoolLogoSetting />
-      <AcademicTerm />
+      <NameSetting />
+      <LogoSetting />
+      <TermSetting />
     </List>
   </div>
 );
 
-export default SchoolSettings;
+export default SchoolSetting;
