+++ src/components/reusables/List/List.tsx
@@ -1,43 +1,34 @@
-import { Children, FC, ReactElement, cloneElement } from 'react';
+import { FC, ReactElement, useEffect, useRef } from 'react';
 import Icon from '~assets/Icons';
 import { BaseBtn } from '../ui/Buttons';
 
-export const List: FC<{ children: ReactElement[] | ReactElement }> = ({
-  children,
-}) => {
-  const value = (Children.toArray(children) as unknown as ReactElement[]).map(
-    (child: ReactElement) => {
-      const el = child.type as unknown as { displayName: string };
+type TList = { children: ReactElement[] | ReactElement };
 
-      if (el.displayName !== 'ListItem') {
-        throw Error('Component only accepts List Item');
-      }
-
-      if (child.props.parent) {
-        throw Error(
-          'ListItem does not accept parent props directly... Leave it omitted'
-        );
-      }
-
-      return cloneElement(child, { parent: true });
-    }
-  );
-
-  return <dl className="divide-y divide-gray-100">{value.map((v) => v)}</dl>;
-};
+export const List: FC<TList> = ({ children }) => (
+  <dl className="divide-y divide-gray-100">{children}</dl>
+);
 
 type TListItem = {
   [key in 'title' | 'description' | 'action']?: ReactElement | string | null;
-} & { parent?: true };
+};
 
 export const ListItem: FC<TListItem> = (props) => {
-  const { parent, title, description, action } = props;
-  if (!parent) {
-    throw Error('Component should be rendered within the List Component');
-  }
+  const element = useRef<HTMLDivElement>(null);
+  const { title, description, action } = props;
+
+  useEffect(() => {
+    element.current?.parentElement?.classList.forEach((className) => {
+      if (!['divide-y', 'divide-gray-100'].includes(className)) {
+        throw Error('Component should be rendered within the List Component');
+      }
+    });
+  }, []);
 
   return (
-    <div className="px-4 py-6 sm:grid items-start relative text-sm sm:grid-cols-3 sm:gap-4 sm:px-0">
+    <div
+      ref={element}
+      className="px-4 py-6 sm:grid items-start relative text-sm sm:grid-cols-3 sm:gap-4 sm:px-0"
+    >
       <dt className="text-sm font-medium relative leading-6 text-gray-900">
         {title}
         {action && (
